bundling validation graph
